back/requirements.txt
alembic==1.10.4
fastapi==0.95.1
postgres==4.0
psycopg2==2.9.6
python-dotenv==1.0.0
uvicorn==0.22.0
